+++ src/components/LoadingSpinner.vue
@@ -1,0 +1,16 @@
+<template>
+  <div>
+    <LoadingSpinnerIcon />
+  </div>
+</template>
+
+<script>
+import LoadingSpinnerIcon from '../assets/loading-spinner.svg';
+
+export default {
+  name: "LoadingSpinner",
+  components: {
+    LoadingSpinnerIcon
+  },
+};
+</script>
